+++ check_python_version.py
@@ -10,14 +10,14 @@ def check():
     print('Running Python', sys.version)
 
     if sys.version_info < MINIMUM_VERSION:
-        print('[!] Python version unsupported. Minimum version required: {}'.format(
+        print('[!] Python version unsupported. Minimum version required: Python {}'.format(
             '.'.join([str(x) for x in MINIMUM_VERSION])
         ))
         sys.exit(1)
 
     # Deprecation.
     if sys.version_info < PREFERRED_MINIMUM_VERSION:
-        print('[WARNING] Python version is deprecated. Preferred version for DSMR-reader is {} or higher.'.format(
+        print('[WARNING] Python version is deprecated. Preferred version for DSMR-reader is Python {} or higher.'.format(
             '.'.join([str(x) for x in PREFERRED_MINIMUM_VERSION])
         ))